Move openConfigurationApp() to new file
This function is now shared between the Alert and Notification delivery
mechanisms.
